# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Tracker API
  version: 1.0.0
  description: JSON REST API for the Medical Tracker Application MVP (api_contract.md).
paths:
...
  /api/v1/examinations/:
    get:
      operationId: examinations_list
      description: |-
        `GET` / `POST /api/v1/examinations/` — only the caller's records; a JSON array.

        The list accepts `search`, `status`, `category`, `ordering` and `time_state`; all
        supplied filters apply together (ADS-FR-028-01 … ADS-FR-032-01). `time_state`
        selects the past, upcoming or overdue collection through examinations/time_state.py.
      parameters:
      - in: query
        name: category
        schema:
          type: integer
        description: A category identifier.
      - in: query
        name: ordering
        schema:
          enum:
          - scheduled_date
          - -scheduled_date
          type: string
          minLength: 1
        description: |-
          * `scheduled_date` - scheduled_date
          * `-scheduled_date` - -scheduled_date
      - in: query
        name: search
        schema:
          type: string
          minLength: 1
          maxLength: 200
        description: Case-insensitive containment on the title.
      - in: query
        name: status
        schema:
          enum:
          - draft
          - planned
          - completed
          - cancelled
          - missed
          type: string
          minLength: 1
        description: |-
          * `draft` - draft
          * `planned` - planned
          * `completed` - completed
          * `cancelled` - cancelled
          * `missed` - missed
      - in: query
        name: time_state
        schema:
          enum:
          - past
          - upcoming
          - overdue
          type: string
          minLength: 1
        description: |-
          A derived collection: past, upcoming or overdue. Distinct from the time_state response field, whose values are upcoming, overdue or null.

          * `past` - past
          * `upcoming` - upcoming
          * `overdue` - overdue
      tags:
      - examinations
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Examination'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
    post:
      operationId: examinations_create
      description: |-
        `GET` / `POST /api/v1/examinations/` — only the caller's records; a JSON array.

        The list accepts `search`, `status`, `category`, `ordering` and `time_state`; all
        supplied filters apply together (ADS-FR-028-01 … ADS-FR-032-01). `time_state`
        selects the past, upcoming or overdue collection through examinations/time_state.py.
      tags:
      - examinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExaminationRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Examination'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
  /api/v1/examinations/{id}/:
    get:
      operationId: examinations_retrieve
      description: |-
        `GET` / `PATCH` / `DELETE /api/v1/examinations/{id}/`.

        The record is looked up only among the caller's own examinations, so a missing
        record and another user's record get the same 404 (ADS-SEC-006-01). `PATCH` is
        partial, but the complete resulting record is validated (ADS-FR-022-01); `PUT` is
        not offered (api_contract.md §12). Deleting cascades to the reminder and the
        recurrence rule and nulls `source_occurrence` on generated occurrences
        (ADS-FR-023-01, ADS-FR-023-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Examination'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
    patch:
      operationId: examinations_partial_update
      description: |-
        `GET` / `PATCH` / `DELETE /api/v1/examinations/{id}/`.

        The record is looked up only among the caller's own examinations, so a missing
        record and another user's record get the same 404 (ADS-SEC-006-01). `PATCH` is
        partial, but the complete resulting record is validated (ADS-FR-022-01); `PUT` is
        not offered (api_contract.md §12). Deleting cascades to the reminder and the
        recurrence rule and nulls `source_occurrence` on generated occurrences
        (ADS-FR-023-01, ADS-FR-023-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedExaminationRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Examination'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
    delete:
      operationId: examinations_destroy
      description: |-
        `GET` / `PATCH` / `DELETE /api/v1/examinations/{id}/`.

        The record is looked up only among the caller's own examinations, so a missing
        record and another user's record get the same 404 (ADS-SEC-006-01). `PATCH` is
        partial, but the complete resulting record is validated (ADS-FR-022-01); `PUT` is
        not offered (api_contract.md §12). Deleting cascades to the reminder and the
        recurrence rule and nulls `source_occurrence` on generated occurrences
        (ADS-FR-023-01, ADS-FR-023-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      security:
      - bearerAuth: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
...
  /api/v1/examinations/{id}/recurrence/:
    get:
      operationId: examinations_recurrence_retrieve
      description: |-
        `GET` / `POST` / `PATCH /api/v1/examinations/{id}/recurrence/` (api_contract.md §15).

        Resolved inside the caller's own examinations (a missing examination or rule is the
        same 404). Creation and updates need a planned examination with a scheduled date
        (ADS-FR-018-01, ADS-FR-039-01); a retained rule stays readable in any status and
        is never changed by a status change (ADS-FR-039-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurrenceRule'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
    post:
      operationId: examinations_recurrence_create
      description: |-
        `GET` / `POST` / `PATCH /api/v1/examinations/{id}/recurrence/` (api_contract.md §15).

        Resolved inside the caller's own examinations (a missing examination or rule is the
        same 404). Creation and updates need a planned examination with a scheduled date
        (ADS-FR-018-01, ADS-FR-039-01); a retained rule stays readable in any status and
        is never changed by a status change (ADS-FR-039-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurrenceRuleRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurrenceRule'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
    patch:
      operationId: examinations_recurrence_partial_update
      description: |-
        `GET` / `POST` / `PATCH /api/v1/examinations/{id}/recurrence/` (api_contract.md §15).

        Resolved inside the caller's own examinations (a missing examination or rule is the
        same 404). Creation and updates need a planned examination with a scheduled date
        (ADS-FR-018-01, ADS-FR-039-01); a retained rule stays readable in any status and
        is never changed by a status change (ADS-FR-039-02).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - examinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedRecurrenceRuleRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurrenceRule'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Not found (api_contract.md §3.3): the record does not exist
            or belongs to another user.'
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
      required:
      - id
      - name
      - slug
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    Examination:
      type: object
      description: |-
        The examination representation and its create/update input (api_contract.md §8).

        Ownership, the source occurrence, the derived time state and the audit instants
        are never accepted from a client (§10); a request that sends one is rejected.
        Every create and update validates the complete resulting record (§8.2).
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: integer
          readOnly: true
        category:
          allOf:
          - $ref: '#/components/schemas/Category'
          readOnly: true
          nullable: true
        title:
          type: string
          maxLength: 200
        medical_specialty:
          type: string
          nullable: true
          maxLength: 200
        scheduled_date:
          type: string
          format: date
          nullable: true
        scheduled_time:
          type: string
          format: time
          nullable: true
        completed_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        location:
          type: string
          nullable: true
          maxLength: 255
        notes:
          type: string
          nullable: true
        source_occurrence:
          type: integer
          readOnly: true
          nullable: true
        time_state:
          readOnly: true
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/TimeStateEnum'
          - $ref: '#/components/schemas/NullEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - category
      - created_at
      - id
      - source_occurrence
      - status
      - time_state
      - title
      - updated_at
      - user_id
    ExaminationRequest:
      type: object
      description: |-
        The examination representation and its create/update input (api_contract.md §8).

        Ownership, the source occurrence, the derived time state and the audit instants
        are never accepted from a client (§10); a request that sends one is rejected.
        Every create and update validates the complete resulting record (§8.2).
      properties:
        category_id:
          type: integer
          writeOnly: true
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        medical_specialty:
          type: string
          nullable: true
          maxLength: 200
        scheduled_date:
          type: string
          format: date
          nullable: true
        scheduled_time:
          type: string
          format: time
          nullable: true
        completed_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        location:
          type: string
          nullable: true
          maxLength: 255
        notes:
          type: string
          nullable: true
      required:
      - status
      - title
...
    NullEnum:
      enum:
      - null
...
    PatchedExaminationRequest:
      type: object
      description: |-
        The examination representation and its create/update input (api_contract.md §8).

        Ownership, the source occurrence, the derived time state and the audit instants
        are never accepted from a client (§10); a request that sends one is rejected.
        Every create and update validates the complete resulting record (§8.2).
      properties:
        category_id:
          type: integer
          writeOnly: true
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        medical_specialty:
          type: string
          nullable: true
          maxLength: 200
        scheduled_date:
          type: string
          format: date
          nullable: true
        scheduled_time:
          type: string
          format: time
          nullable: true
        completed_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        location:
          type: string
          nullable: true
          maxLength: 255
        notes:
          type: string
          nullable: true
    PatchedRecurrenceRuleRequest:
      type: object
      description: |-
        The recurrence representation (api_contract.md §15).

        `next_due_date` is derived from the examination's current scheduled date with
        calendar-month arithmetic, and is null while there is none.
      properties:
        interval:
          $ref: '#/components/schemas/IntervalEnum'
...
    RecurrenceRule:
      type: object
      description: |-
        The recurrence representation (api_contract.md §15).

        `next_due_date` is derived from the examination's current scheduled date with
        calendar-month arithmetic, and is null while there is none.
      properties:
        id:
          type: integer
          readOnly: true
        examination:
          type: integer
          readOnly: true
        interval:
          $ref: '#/components/schemas/IntervalEnum'
        next_due_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - examination
      - id
      - interval
      - next_due_date
      - updated_at
    RecurrenceRuleRequest:
      type: object
      description: |-
        The recurrence representation (api_contract.md §15).

        `next_due_date` is derived from the examination's current scheduled date with
        calendar-month arithmetic, and is null while there is none.
      properties:
        interval:
          $ref: '#/components/schemas/IntervalEnum'
      required:
      - interval
...
    TimeStateEnum:
      enum:
      - upcoming
      - overdue
      type: string
      description: |-
        * `upcoming` - upcoming
        * `overdue` - overdue
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT